requirements.txt
beautifulsoup4==4.10.0
pymongo==3.12.1
requests==2.27.1
PyYAML==6.0
schedule==1.1.0
